invoice-processing-pipeline/reviewer/requirements.txt
flask>=2.3.3
google-auth>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
gunicorn>=21.2.0
redis==5.0.0
